Rewrite model test to class. Add conftest file.
